+++ src/main/java/com/estsoft/guesshangeul/admin/entity/BoardManagerApply.java
@@ -1,14 +1,18 @@
-package com.estsoft.guesshangeul.entity;
+package com.estsoft.guesshangeul.admin.entity;
 
 import java.time.LocalDateTime;
 
 import org.hibernate.annotations.ColumnDefault;
+
+import com.estsoft.guesshangeul.users.entity.Users;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -18,23 +22,21 @@ import lombok.Setter;
 @Setter
 @Entity
 @NoArgsConstructor
-@Table(name = "quiz_board")
-public class QuizBoard {
+@Table(name = "board_manager_apply")
+public class BoardManagerApply {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "quiz_board_id", unique = true, nullable = false)
+	@Column(name = "board_manager_apply_id", unique = true, nullable = false)
 	private Long id;
 
-	@Column(nullable = false)
-	private String title;
+	@ManyToOne
+	@JoinColumn(name = "user_id")
+	private Users users;
 
 	@Column(nullable = false)
-	private Long userId;
+	@ColumnDefault("0")
+	private int status;
 
 	@Column(name = "created_at")
 	private LocalDateTime createdAt;
-
-	@Column(name = "is_deleted", nullable = false)
-	@ColumnDefault("false")
-	private Boolean isDeleted;
 }
