+++ src/main/java/com/estsoft/guesshangeul/board/entity/GeneralBoard.java
@@ -1,14 +1,14 @@
-package com.estsoft.guesshangeul.entity;
+package com.estsoft.guesshangeul.board.entity;
 
 import java.time.LocalDateTime;
+
+import org.hibernate.annotations.ColumnDefault;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -16,26 +16,22 @@ import lombok.Setter;
 
 @Getter
 @Setter
-@NoArgsConstructor
 @Entity
-@Table(name = "general_comment")
-public class GeneralComment {
+@NoArgsConstructor
+@Table(name = "general_board")
+public class GeneralBoard {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "general_comment_id", unique = true, nullable = false)
+	@Column(name = "general_board_id", unique = true, nullable = false)
 	private Long id;
 
-	@ManyToOne
-	@JoinColumn(name = "user_id")
-	private Users users;
-
-	@ManyToOne
-	@JoinColumn(name = "general_post_id")
-	private GeneralPost post;
-
 	@Column(nullable = false)
-	private String content;
+	private String title;
 
-	@Column(nullable = false)
+	@Column(name = "created_at")
 	private LocalDateTime createdAt;
+
+	@Column(name = "is_deleted", nullable = false)
+	@ColumnDefault("false")
+	private Boolean isDeleted;
 }
