+++ src/main/java/com/estsoft/guesshangeul/board/entity/QuizBoard.java
@@ -1,4 +1,4 @@
-package com.estsoft.guesshangeul.entity;
+package com.estsoft.guesshangeul.board.entity;
 
 import java.time.LocalDateTime;
 
@@ -9,8 +9,6 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -20,21 +18,23 @@ import lombok.Setter;
 @Setter
 @Entity
 @NoArgsConstructor
-@Table(name = "board_manager_apply")
-public class BoardManagerApply {
+@Table(name = "quiz_board")
+public class QuizBoard {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "board_manager_apply_id", unique = true, nullable = false)
+	@Column(name = "quiz_board_id", unique = true, nullable = false)
 	private Long id;
 
-	@ManyToOne
-	@JoinColumn(name = "user_id")
-	private Users users;
+	@Column(nullable = false)
+	private String title;
 
 	@Column(nullable = false)
-	@ColumnDefault("0")
-	private int status;
+	private Long userId;
 
 	@Column(name = "created_at")
 	private LocalDateTime createdAt;
+
+	@Column(name = "is_deleted", nullable = false)
+	@ColumnDefault("false")
+	private Boolean isDeleted;
 }
