+++ src/main/java/com/estsoft/guesshangeul/post/entity/GeneralPost.java
@@ -1,0 +1,59 @@
+package com.estsoft.guesshangeul.post.entity;
+
+import java.time.LocalDateTime;
+
+import org.hibernate.annotations.ColumnDefault;
+
+import com.estsoft.guesshangeul.board.entity.GeneralBoard;
+import com.estsoft.guesshangeul.users.entity.Users;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@Entity
+@NoArgsConstructor
+@Table(name = "general_post")
+public class GeneralPost {
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	@Column(name = "general_post_id", unique = true, nullable = false)
+	private Long id;
+
+	@ManyToOne
+	@JoinColumn(name = "user_id")
+	private Users users;
+
+	@ManyToOne
+	@JoinColumn(name = "general_board_id")
+	private GeneralBoard generalBoard;
+
+	@Column(nullable = false)
+	private String title;
+
+	@Column(nullable = false)
+	private String content;
+
+	@Column(name = "is_hidden", nullable = false)
+	@ColumnDefault("false")
+	private boolean isHidden;
+
+	@Column(nullable = false)
+	private Long view;
+
+	@Column(name = "created_at")
+	private LocalDateTime createdAt;
+
+	@Column(name = "updated_at")
+	private LocalDateTime updatedAt;
+}
