+++ src/main/java/com/estsoft/guesshangeul/post/entity/PostUser.java
@@ -1,4 +1,6 @@
-package com.estsoft.guesshangeul.entity;
+package com.estsoft.guesshangeul.post.entity;
+
+import com.estsoft.guesshangeul.users.entity.Users;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.ManyToOne;
