+++ src/main/java/com/estsoft/guesshangeul/users/controller/UsersController.java
@@ -1,0 +1,8 @@
+package com.estsoft.guesshangeul.users.controller;
+
+import org.springframework.stereotype.Controller;
+
+@Controller
+public class UsersController {
+
+}
