+++ src/main/java/com/estsoft/guesshangeul/users/entity/Users.java
@@ -1,4 +1,4 @@
-package com.estsoft.guesshangeul.entity;
+package com.estsoft.guesshangeul.users.entity;
 
 import java.time.LocalDateTime;
